Keep launch-cover contract within structural budget
